+++ src/Ontogony.Security/OntogonyServiceIdentityHeaders.cs
@@ -6,6 +6,7 @@ namespace Ontogony.Security;
 public static class OntogonyServiceIdentityHeaders
 {
     public const string ServiceId = "X-Ontogony-Service-Id";
+    public const string KeyId = "X-Ontogony-Service-Key-Id";
     public const string Timestamp = "X-Ontogony-Service-Timestamp";
     public const string Nonce = "X-Ontogony-Service-Nonce";
     public const string BodyHash = "X-Ontogony-Service-Body-Hash";